src/main/java/edu/ucsb/cs156/example/services/GrantedAuthoritiesService.java
package edu.ucsb.cs156.example.services;

import java.util.Collection;

import org.springframework.security.core.Authentication;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.context.SecurityContext;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Service;

import lombok.extern.slf4j.Slf4j;

@Slf4j
@Service("grantedAuthorities")
public class GrantedAuthoritiesService {

    public Collection<? extends GrantedAuthority> getGrantedAuthorities() {
        SecurityContext securityContext = SecurityContextHolder.getContext();
        Authentication authentication = securityContext.getAuthentication();
        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
        log.info("authorities={}", authorities);
        return authorities;
    }

}
